big push

+++ src/App.js
@@ -1,10 +1,12 @@
+import { Box, Container } from '@mui/material';
+import { useState, useEffect } from 'react';
 import './App.css';
 import Header from './Header';
 import Editor from './Editor';
-import Avatar from './Avatar';
+import Customizer from './Customizer';
 import Wallet from './Wallet';
-import { Box, Container } from '@mui/material';
-import { useState, useEffect } from 'react';
+import initialShop from './InitialShop';
+import initialCloset from './InitialCloset';
 
 function App() {
   // populate local storage with an initial list
@@ -41,14 +43,31 @@ function App() {
     localStorage.setItem("day", JSON.stringify(new Date().getDate()));
   }
   if (!localStorage.getItem("dailyMoney")) {
-    localStorage.setItem("dailyMoney", JSON.stringify(0));
+    if (localStorage.getItem("listManager")) { // presumably already mounted once--why are you deleting the daily limit hmm?
+      localStorage.setItem("dailyMoney", JSON.stringify(100));
+    } else {
+      localStorage.setItem("dailyMoney", JSON.stringify(0));
+    }
+  }
+  if (!localStorage.getItem("wearing")) {
+    localStorage.setItem("wearing", JSON.stringify([initialCloset[0]]));
+  }
+  if (!localStorage.getItem("closet")) {
+    localStorage.setItem("closet", JSON.stringify(initialCloset));
+  }
+  if (!localStorage.getItem("shop")) {
+    localStorage.setItem("shop", JSON.stringify(initialShop));
   }
 
   let listManager = JSON.parse(localStorage.getItem("listManager"));
-  const initCompletedTasks = JSON.parse(localStorage.getItem("completedTasks"));
   const initList = JSON.parse(localStorage.getItem("list"));
+  const initCompletedTasks = JSON.parse(localStorage.getItem("completedTasks"));
   const initMoney = parseInt(localStorage.getItem("wallet"));
   const initDailyMoney = parseInt(localStorage.getItem("dailyMoney"));
+  // customization stuffs
+  const initWearing = JSON.parse(localStorage.getItem("wearing"));
+  const initCloset = JSON.parse(localStorage.getItem("closet"));
+  const initShop = JSON.parse(localStorage.getItem("shop"));
 
   /* --------------------════ ⋆★⋆ ════-------------------- */
 
@@ -65,11 +84,23 @@ function App() {
   const [isEmptyTask, setIsEmptyTask] = useState(false);
   const [completedOpen, setCompletedOpen] = useState(false);
   const [maxTasksRestored, setMaxTasksRestored] = useState(false);
+  const [importOpen, setImportOpen] = useState(false);
+  const [importedFile, setImportedFile] = useState(null);
+  const [invalidFileOpen, setInvalidFileOpen] = useState(false);
+  const [tooLongFileOpen, setTooLongFileOpen] = useState(false);
+  const [repNameFileOpen, setRepNameFileOpen] = useState(false);
 
   const [money, setMoney] = useState(initMoney);
   const [dailyMoney, setDailyMoney] = useState(initDailyMoney);
   const moneyCap = 10000;
   const dailyMoneyCap = 100;
+
+  const [wearing, setWearing] = useState(initWearing);
+  const [closet, setCloset] = useState(initCloset);
+  const [shop, setShop] = useState(initShop);
+  const [currentItem, setCurrentItem] = useState(null);
+
+  const [lastCheckTime, setLastCheckTime] = useState(0);
 
   /* --------------------════ ⋆★⋆ ════-------------------- */
 
@@ -82,6 +113,24 @@ function App() {
       localStorage.setItem("dailyMoney", JSON.stringify(0));
     }
   }, []);
+
+  useEffect(() => {
+    if (closet.length + shop.length !== initialShop.length) { // if there's a new item added
+      console.log("new item in shop woop de doo");
+      initialShop.forEach(item => {
+        const foundInCloset = closet.find((element) => element.filename === item.filename);
+        const foundInShop = shop.find((element) => element.filename === item.filename);
+        if (!foundInCloset && !foundInShop) {
+          const newShop = [
+            ...shop,
+            item
+          ]
+          setShop(newShop);
+          localStorage.setItem('shop', JSON.stringify(newShop));
+        }
+      });
+    }
+  }, [closet, shop]);
 
   /* --------------------════ ⋆★⋆ ════-------------------- */
 
@@ -99,7 +148,8 @@ function App() {
     const name = listManager.find((element) => element.title === newName);
     if (name) { // if the name already exists
       setNameExists(true);
-    } else {
+    }
+    else {
       setNameExists(false);
       setEditOpen(false); // for closing the dialog
       const index = listManager.findIndex((element) => element.title === list.title);
@@ -125,7 +175,8 @@ function App() {
       setTimeout(() => {
         setNoAddListOpen(false);
       }, 3000);
-    } else {
+    }
+    else {
       setAddOpen(true);
     }
   }
@@ -133,7 +184,8 @@ function App() {
     const name = listManager.find((element) => element.title === newName);
     if (name) {
       setNameExists(true);
-    } else {
+    }
+    else {
       setNameExists(false);
       setAddOpen(false);
       const newList = {
@@ -146,17 +198,30 @@ function App() {
       localStorage.setItem("listManager", JSON.stringify(listManager));
     }
   }
-
-  const handleDeleteOpen = (bool) => {
-    setDeleteOpen(bool);
+  const handleAddExistingList = (newList) => {
+    const name = listManager.find((element) => element.title === newList.title);
+    if (!name) {
+      setList(newList);
+      localStorage.setItem("list", JSON.stringify(newList));
+      listManager.push(newList);
+      localStorage.setItem("listManager", JSON.stringify(listManager));
+    }
+    else {
+      setRepNameFileOpen(true);
+      setTimeout(() => {
+        setRepNameFileOpen(false);
+      }, 3000);
+    }
   }
+
   const handleDeleteListButton = () => {
     if (listManager.length <= 1) {
       setNoDeleteListOpen(true);
       setTimeout(() => {
         setNoDeleteListOpen(false);
       }, 3000);
-    } else {
+    }
+    else {
       setDeleteOpen(true);
     }
   }
@@ -167,7 +232,8 @@ function App() {
     let newList = null;
     if (index === 0) {
         newList = listManager[index + 1];
-    } else {
+    }
+    else {
         newList = listManager[index - 1];
     }
     setList(newList);
@@ -201,12 +267,14 @@ function App() {
       setTimeout(() => {
         setIsEmptyTask(false);
       }, 3000);
-    } else if (list.data.length >= 15) {
+    }
+    else if (list.data.length >= 15) {
       setMaxTasksReached(true);
       setTimeout(() => {
         setMaxTasksReached(false);
       }, 3000);
-    } else {
+    }
+    else {
       const listIndex = listManager.findIndex((element) => element.title === list.title);
 
       let newList = {...list};
@@ -240,55 +308,73 @@ function App() {
   }
 
   const handleCheckTask = (e, index, page, itemsPerPage) => {
-    const newIndex = index + (itemsPerPage * (page - 1));
-    const listIndex = listManager.findIndex((element) => element.title === list.title);
+    const currentTime = Date.now();
+    if (currentTime - lastCheckTime > 300) { // prevents you from checking too fast (so setInterval can do its thing)
+      setLastCheckTime(currentTime);
 
-    let newList = {...list};
-    let newData = [...newList.data];
-    newData[newIndex] = {
-      ...newData[newIndex],
-      complete: e.target.checked
-    };
-    newList.data = newData;
-    
-    setList(newList);
-    localStorage.setItem("list", JSON.stringify(newList));
-    listManager[listIndex] = newList;
-    localStorage.setItem("listManager", JSON.stringify(listManager));
+      const newIndex = index + (itemsPerPage * (page - 1));
+      const listIndex = listManager.findIndex((element) => element.title === list.title);
 
-    // money increment
-    let newMoney = money;
-    let newDailyMoney = dailyMoney;
-    const completedIndex = completedTasks.findIndex((element) => element.name === newData[newIndex].name)
-    let newCompletedTasks = [...completedTasks];
-    if (e.target.checked) {
-      newCompletedTasks = [
-        ...newCompletedTasks,
-        newData[newIndex]
-      ];
-      if (newCompletedTasks.length > 15) { // auto clears old tasks
-        newCompletedTasks.shift();
-      }
+      let newList = { ...list };
+      let newData = [...newList.data];
+      newData[newIndex] = {
+        ...newData[newIndex],
+        complete: e.target.checked
+      };
+      newList.data = newData;
 
-      if (money < moneyCap && dailyMoney < dailyMoneyCap) { // money completely freezes when you hit either cap
-        newMoney = money + 10;
-        setMoney(newMoney); // could also be "setMoney(money + 10)" but i need newMoney for localStorage
-        newDailyMoney = dailyMoney + 10;
-        setDailyMoney(newDailyMoney);
+      setList(newList);
+      localStorage.setItem("list", JSON.stringify(newList));
+      listManager[listIndex] = newList;
+      localStorage.setItem("listManager", JSON.stringify(listManager));
+
+      // money increment
+      let oldMoney = money
+      let newMoney = money;
+      let newDailyMoney = dailyMoney;
+      const completedIndex = completedTasks.findIndex((element) => element.name === newData[newIndex].name)
+      let newCompletedTasks = [...completedTasks];
+
+      if (e.target.checked) {
+        newCompletedTasks = [
+          ...newCompletedTasks,
+          newData[newIndex]
+        ];
+        if (newCompletedTasks.length > 15) { // auto clears old tasks
+          newCompletedTasks.shift();
+        }
+
+        if (money < moneyCap && dailyMoney < dailyMoneyCap) { // money completely freezes when you hit either cap
+          const counter = setInterval(() => {
+            newMoney += 1;
+            setMoney(newMoney);
+            if (newMoney === oldMoney + 10) {
+              clearInterval(counter);
+            }
+          }, 20);
+          newDailyMoney = dailyMoney + 10;
+          setDailyMoney(newDailyMoney);
+        }
       }
-    } else {
-      newCompletedTasks.splice(completedIndex, 1);
-      if (money < moneyCap && dailyMoney < dailyMoneyCap) {
-        newMoney = money - 10;
-        setMoney(newMoney);
-        newDailyMoney = dailyMoney - 10;
-        setDailyMoney(newDailyMoney);
+      else {
+        newCompletedTasks.splice(completedIndex, 1);
+        if (money < moneyCap && dailyMoney < dailyMoneyCap) {
+          const counter = setInterval(() => {
+            newMoney -= 1;
+            setMoney(newMoney);
+            if (newMoney === oldMoney - 10) {
+              clearInterval(counter);
+            }
+          }, 20);
+          newDailyMoney = dailyMoney - 10;
+          setDailyMoney(newDailyMoney);
+        }
       }
+      setCompletedTasks(newCompletedTasks);
+
+      localStorage.setItem("wallet", JSON.stringify(newMoney));
+      localStorage.setItem("dailyMoney", JSON.stringify(newDailyMoney));
     }
-    setCompletedTasks(newCompletedTasks);
-    
-    localStorage.setItem("wallet", JSON.stringify(newMoney));
-    localStorage.setItem("dailyMoney", JSON.stringify(newDailyMoney));
   }
 
   const handleClearList = () => {
@@ -320,7 +406,8 @@ function App() {
     
     if (list.data.length >= 15) {
       setMaxTasksRestored(true);
-    } else {
+    }
+    else {
       setMaxTasksRestored(false);
 
       const listIndex = listManager.findIndex((element) => element.title === list.title);
@@ -345,53 +432,221 @@ function App() {
     }
   }
 
+  const handleImport = (file, replace) => {
+    const reader = new FileReader();
+
+    reader.onload = (event) => {
+        const fileContents = event.target.result; // file contents as text
+        let fileObject = null;
+        try { // try-catch block to account for invalid files (will throw error on JSON.parse())
+          fileObject = JSON.parse(fileContents);
+
+          if (replace) {
+            if (fileObject.length <= 10) { // shouldn't be possible when exporting but in case someone inputs a faulty file...
+              setList(fileObject[0]);
+              localStorage.setItem("list", JSON.stringify(fileObject[0]));
+              listManager = fileObject;
+              localStorage.setItem("listManager", JSON.stringify(listManager));
+            }
+            else {
+              setTooLongFileOpen(true);
+              setTimeout(() => {
+                setTooLongFileOpen(false);
+              }, 3000);
+            }
+          } 
+          else {
+            if (fileObject.length + listManager.length <= 10) {
+                for (let i = 0; i < fileObject.length; i++) {
+                    handleAddExistingList(fileObject[i]);
+                }
+            }
+            else {
+              setTooLongFileOpen(true);
+              setTimeout(() => {
+                setTooLongFileOpen(false);
+              }, 3000);
+            }
+          }
+        }
+        catch (e) {
+          console.log(e.stack);
+          setInvalidFileOpen(true);
+          setTimeout(() => {
+            setInvalidFileOpen(false);
+          }, 3000);
+        }
+    };
+
+    reader.readAsText(file);
+
+    setImportOpen(false);
+    setImportedFile(null);
+  }
+
+  const handleExport = () => {
+    const stringifiedLists = JSON.stringify(listManager, null, 4); // 4 for indentation
+    const file = new Blob([stringifiedLists], { type: 'text/plain' });
+
+    // anchor link
+    const element = document.createElement("a");
+    element.href = URL.createObjectURL(file);
+    element.download = "list.txt";
+
+    // simulate link click
+    document.body.appendChild(element); // required for this to work in FireFox
+    element.click();
+    element.parentNode.removeChild(element); // reset
+  }
+
+  const handleClickShopItem = (filename) => {
+    const item = shop.find((element) => element.filename === filename);
+    if (currentItem && filename === currentItem.filename) { // if deselecting
+      const newWearing = JSON.parse(localStorage.getItem("wearing")); // go back to original outfit
+      setWearing(newWearing);
+      setCurrentItem(null);
+    } else {
+      if (wearing.some((element) => element.category === item.category) || item.category === 'character') { // if exists, replace
+        const index = wearing.findIndex((element) => element.category === item.category);
+        let newWearing = [...wearing];
+        newWearing[index] = item;
+        setWearing(newWearing);
+      } else { // if doesn't exist, add
+        const newWearing = [
+          ...wearing,
+          item
+        ]
+        setWearing(newWearing);
+      }
+      setCurrentItem(item);
+    }
+  }
+  const handleBuy = () => {
+    if (money >= currentItem.price) {
+      // before, i did findIndex() + splice() but this seems to work just fine
+      const newShop = shop.filter((item) => item.filename !== currentItem.filename);
+      setShop(newShop);
+      const newCloset = [
+        ...closet,
+        currentItem
+      ]
+      setCloset(newCloset);
+
+      const newMoney = money - currentItem.price;
+      setMoney(newMoney);
+
+      const newWearing = wearing.filter((item) => item.filename !== currentItem.filename);
+      setWearing(newWearing);
+      setCurrentItem(null);
+
+      localStorage.setItem("shop", JSON.stringify(newShop));
+      localStorage.setItem("closet", JSON.stringify(newCloset));
+      localStorage.setItem("wallet", JSON.stringify(newMoney));
+    }
+  }
+  const handleClickClosetItem = (filename) => {
+    let newWearing;
+    const item = closet.find((element) => element.filename === filename);
+    if (item.category === 'character') { // does not account for deselecting bc user must have one character
+      newWearing = [...wearing];
+      newWearing[0] = item;
+      setWearing(newWearing);
+      setCurrentItem(item);
+    } else {
+      if (wearing.some((element) => element.filename === item.filename)) { // if already wearing
+        newWearing = wearing.filter((element) => element.filename !== item.filename);
+        setWearing(newWearing);
+        setCurrentItem(null);
+      } else {
+        if (wearing.some((element) => element.category === item.category)) {
+          const index = wearing.findIndex((element) => element.category === item.category);
+          newWearing = [...wearing];
+          newWearing[index] = item;
+          setWearing(newWearing);
+        } else {
+          newWearing = [
+            ...wearing,
+            item
+          ]
+          setWearing(newWearing);
+        }
+        setCurrentItem(item);
+      }
+    }
+    localStorage.setItem("wearing", JSON.stringify(newWearing));
+  }
+
   /* --------------------════ ⋆★⋆ ════-------------------- */
 
   return (
     <Container className="App" maxWidth='lg'>
-      <Box sx={{ minWidth: 910, display: 'flex', justifyContent: 'space-between' }}>
-        <Header text="Today's Tasks"/>
-        <Wallet value={money}/>
+      <Box sx={{ width: 1160, display: 'flex', justifyContent: 'space-between' }}>
+        <Header text="CheckMate" />
+        <Wallet value={money} />
       </Box>
       <Box sx={{ // outer brown box
         bgcolor: '#837370',
         borderRadius: '32px',
         height: 580, // height is hard-coded because i do not want it to stretch vertically
-        minWidth: 850,
+        width: 1100, // used to be minwidth but i don't want to deal w resizing anymore...
         marginTop: '16px',
         padding: '30px',
         display: 'flex'
       }}>
         <Editor
           handleClickMenu={handleClickMenu}
+
           editOpen={editOpen}
           handleEditOpen={handleEditOpen}
           nameExists={nameExists}
           handleEditList={handleEditList}
+
           addOpen={addOpen}
           noAddListOpen={noAddListOpen}
           handleAddOpen={handleAddOpen}
           handleAddListButton={handleAddListButton}
           handleAddListDialog={handleAddListDialog}
+          handleAddExistingList={handleAddExistingList}
+
           deleteOpen={deleteOpen}
           noDeleteListOpen={noDeleteListOpen}
-          handleDeleteOpen={handleDeleteOpen}
+          setDeleteOpen={setDeleteOpen}
           handleDeleteListButton={handleDeleteListButton}
           handleDeleteListDialog={handleDeleteListDialog}
+
           handleEditTask={handleEditTask}
           maxTasksReached={maxTasksReached}
           isEmptyTask={isEmptyTask}
           handleAddTask={handleAddTask}
           handleDeleteTask={handleDeleteTask}
           handleCheckTask={handleCheckTask}
+
           handleClearList={handleClearList}
           completedOpen={completedOpen}
           handleCompletedOpen={handleCompletedOpen}
           handleEmptyTrash={handleEmptyTrash}
           handleRestoreTask={handleRestoreTask}
           maxTasksRestored={maxTasksRestored}
+
+          importOpen={importOpen}
+          setImportOpen={setImportOpen}
+          handleImport={handleImport}
+          importedFile={importedFile}
+          setImportedFile={setImportedFile}
+          handleExport={handleExport}
+          invalidFileOpen={invalidFileOpen}
+          tooLongFileOpen={tooLongFileOpen}
+          repNameFileOpen={repNameFileOpen}
         />
-        <Avatar/>
+        <Customizer
+          currentItem={currentItem}
+          setCurrentItem={setCurrentItem}
+          handleClickShopItem={handleClickShopItem}
+          wearing={wearing}
+          setWearing={setWearing}
+          handleBuy={handleBuy}
+          handleClickClosetItem={handleClickClosetItem}
+        />
       </Box>
     </Container>
   );
